Fix test 'Approx' compilation error and create ActivationFunction.hpp
